refactor(media): extract PlexSettingsPage result display components [tb-367]
Extract 4 result display components (SyncResultDisplay,
WatchlistSyncResultDisplay, DiscoverSyncResultDisplay,
WatchHistorySyncResultDisplay) and shared types into dedicated files.
PlexSettingsPage.tsx reduced from 602 → 207 LOC. Section components
now import types from shared types.ts instead of declaring locally.

Closes #1714
